requirements.txt
# Core Framework & Orchestration
langchain>=0.1.14
langgraph>=0.0.26
langchain-openai>=0.1.0
langchain-anthropic>=0.1.4
pydantic>=2.6.0
python-dotenv>=1.0.1
pyyaml>=6.0.1

# API Server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Data Processing & RAG (Retrieval-Augmented Generation)
chromadb>=0.4.22
sentence-transformers>=2.2.2
langchain-text-splitters>=0.0.2
numpy>=1.24.0
scikit-learn>=1.4.0  # for isotonic regression or confidence scoring

# Fine-tuning & Model Serving
transformers>=4.41.0
datasets>=2.18.0
torch>=2.2.0
accelerate>=0.29.3
bitsandbytes>=0.42.0
peft>=0.10.0

# Observability
langsmith>=0.1.21

# --- Development & Notebooks ---
jupyterlab>=4.1.0
notebook>=7.2.0
